Skeleton swordman and Skeleton archer added
Skeleton sword man patrols but chases player when he is near and then attacks the player with a sword when he is in range.
Skeleton archer patrols but attacks player when he is in range but when player is near the archer he runs away unless he is against a wall(no where left to run), then he stops running away and starts attacking the player.